A36926-000
Testing - Via test Hardware Change only -
Vprog is stored before each pulse.  This should be made permanent in
respin

Updated to CAN version 0x17.  The HV lambda on/off is now controlled by
control bit in sync message from the ECB instead of looking at if it is
powered or not